added print

+++ Inputs.py
@@ -133,7 +133,7 @@ def get_all_test_data(im_list, la_list):
     for im_filename, la_filename in zip(im_list, la_list):
         print(im_filename)
         print(la_filename)
-        im = np.array(skimage.io.imread(im_filename), np.float32)
+        im = np.array(skimage.io.imread(im_filename), np.int32)
         im = resize(im, (360, 480))
         im = im[np.newaxis]
         la = skimage.io.imread(la_filename)
